DataLayer now accepts float for the label

+++ include/caffe/data_layers.hpp
@@ -237,7 +237,7 @@ class ImageDataLayer : public BasePrefetchingDataLayer<Dtype> {
   virtual void ShuffleImages();
   virtual void InternalThreadEntry();
 
-  vector<std::pair<std::string, int> > lines_;
+  vector<std::pair<std::string, double> > lines_;
   int lines_id_;
 };
